requirements.txt
paddlepaddle>=2.4.0
pydot
loguru
appdirs
